Updated Recurring Task overlap
overlapsWith() prints out the date that causes the conflict. added a test block of code that tests a new R! task against all occurrences of the existing task. Previous commit only added the ability to check against one task's instances, but the block of code will be moved to the overlap method for R! tasks