Project passes all Scenarios
Finished the validation stuff for all kinds of tasks. The 2 scenarios from Canvas are know met and prove the majority of our system works. Antitask was adapted to fit the format of the other tasks and did not make use of its states. removed the old overlap method for the new one.